Ftr fseguel/1209/implement collision

+++ core/src/main/java/org/auSpherical/auFight/placeholders/Damageable.java
@@ -1,5 +1,5 @@
 package org.auSpherical.auFight.placeholders;
 
 public interface Damageable {
-    void receiveDamage(int damage);
+    void receiveDamage(float damage);
 }
